+++ src/raytracers/SequentialRayTracer.cpp
@@ -140,7 +140,7 @@ namespace RayTracing {
                 }
             }
             iteration++;
-            progress = iteration / (static_cast<double>(rays.size()));
+            progress = (double) iteration / (static_cast<double>(rays.size()));
             if (iteration % 1000 == 0) {
                 std::cout << "\r" << iteration << "/" << rays.size() << " rays traced (" << (progress * 100.0) << "%)"
                         << std::flush;
